ExistentialTherapyAutomation/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
rich>=13.7.0
pytz>=2023.3
